Refactor results and recipe detail tests
